requirements.txt
MouseInfo==0.1.3
Pillow==9.0.1
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
pytesseract==0.3.7
PyTweening==1.0.3
